+++ src/pages/home/_ticker.js
@@ -1,164 +1,12 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import styled from 'styled-components'
-import PropTypes from 'prop-types'
+import Tick from './_tick'
 import { BinarySocketBase } from 'common/websocket/socket_base'
 import AutoCarousel from 'components/elements/auto-carousel'
-import { Text } from 'components/elements/typography.js'
-import Loader from 'components/elements/dot_loader.js'
-import MovementGreen from 'images/svg/price-movement-green.svg'
-import MovementRed from 'images/svg/price-movement-red.svg'
 
-const Divider = styled.div`
-    width: 1px;
-    height: 2.6rem;
-    background-color: var(--color-grey-6);
-`
-const TickWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    padding: 2.7rem 0;
-    width: 38rem;
-`
-const StyledText = styled(Text)`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    margin: 0 auto;
-    font-size: var(--text-size-sm);
-`
-const Qoute = styled.span`
-    font-weight: bold;
-    text-decoration: none;
-    padding: 0 5px;
-`
 const CarouselWapper = styled.div`
     box-shadow: 0 16px 20px 0 rgba(0, 0, 0, 0.1);
 `
-class Tick extends React.PureComponent {
-    state = {
-        quote: null,
-        movement: null,
-    }
-
-    subscribe = symbol => {
-        BinarySocketBase.send(
-            {
-                ticks: symbol,
-                subscribe: 1,
-            },
-            { callback: this.onSubscribe },
-        )
-    }
-
-    getHistory = symbol => {
-        BinarySocketBase.send(
-            {
-                ticks_history: symbol,
-                end: 'latest',
-                adjust_start_time: 1,
-                count: 1,
-            },
-            { callback: this.onTicksHistory },
-        )
-    }
-
-    reformatQuote(number) {
-        // calculate that how many decimal each quote should have, based on pip value
-        return number.toFixed(
-            Math.max(
-                (this.props.pip.toString().split('.')[1] || []).length,
-                (number.toString().split('.')[1] || []).length,
-            ),
-        )
-    }
-
-    onTicksHistory = response => {
-        if (response.error) {
-            this.setState({
-                quote: null,
-                movement: null,
-            })
-        } else {
-            this.setState({
-                quote: this.reformatQuote(response.history.prices[0]),
-                movement: null,
-            })
-        }
-    }
-
-    onSubscribe = response => {
-        if (response.error) {
-            this.setState({
-                quote: null,
-                movement: null,
-            })
-        } else {
-            const quote = this.reformatQuote(response.tick.quote)
-            if (this.state.quote > response.tick.quote) {
-                this.setState({
-                    quote,
-                    movement: MovementRed,
-                })
-            } else if (this.state.quote < response.tick.quote) {
-                this.setState({
-                    quote,
-                    movement: MovementGreen,
-                })
-            } else if (this.state.quote === response.tick.quote) {
-                this.setState({
-                    movement: null,
-                })
-            } else {
-                this.setState({
-                    quote,
-                    movement: null,
-                })
-            }
-        }
-    }
-
-    componentDidMount() {
-        const { symbol, is_exchange_open } = this.props
-
-        if (is_exchange_open) {
-            this.subscribe(symbol)
-        } else {
-            this.getHistory(symbol)
-        }
-    }
-
-    render() {
-        const Movement = this.state.movement
-        return (
-            <TickWrapper>
-                <StyledText>
-                    <Qoute>
-                        <span style={{ fontWeight: 'normal' }}>
-                            {this.props.display_name}:{' '}
-                        </span>
-                        {this.state.quote === null ? (
-                            <Loader />
-                        ) : (
-                            this.state.quote
-                        )}{' '}
-                    </Qoute>
-                    <span style={{ width: '12px', display: 'block' }}>
-                        {Movement === null ? null : <Movement />}
-                    </span>
-                </StyledText>
-                <Divider />
-            </TickWrapper>
-        )
-    }
-}
-
-Tick.propTypes = {
-    display_name: PropTypes.string,
-    is_exchange_open: PropTypes.bool,
-    pip: PropTypes.number,
-    symbol: PropTypes.string,
-}
 
 //  Fisher-Yates shuffle
 function shuffle(array) {
@@ -197,55 +45,116 @@ const getTickerMarkets = active_symbols => {
 
     return [...volidx, ...forex]
 }
-class Ticker extends React.Component {
-    state = {
-        markets: [],
-    }
-    onActiveSymbolReceive = response => {
-        const markets = getTickerMarkets(response.active_symbols)
+const SeparateSymbols = symbols_array => {
+    const open_symbols = []
+    const close_symbols = []
 
-        this.setState({
-            markets,
-        })
+    symbols_array.forEach(symbol =>
+        symbol.exchange_is_open === 1
+            ? open_symbols.push(symbol.symbol)
+            : close_symbols.push(symbol.symbol),
+    )
+    return {
+        selected_symbols: symbols_array,
+        close_symbols: close_symbols,
+        open_symbols: open_symbols,
     }
-    componentDidMount() {
+}
+const Ticker = () => {
+    const [markets, setMarkets] = useState({
+        selected_symbols: [],
+        close_symbols: [],
+        open_symbols: [],
+    })
+    const [quotes, setQuotes] = useState()
+
+    useEffect(() => {
         BinarySocketBase.send(
             {
                 active_symbols: 'brief',
                 product_type: 'basic',
             },
-            { callback: this.onActiveSymbolReceive },
+            { callback: onMarketReceive },
+        )
+    }, [])
+
+    const initiateQuotes = incoming_symbols => {
+        const initiatedQuotes = {}
+        incoming_symbols.forEach(
+            symbol => (initiatedQuotes[symbol.symbol] = null),
+        )
+        return initiatedQuotes
+    }
+
+    onMarketReceive = response => {
+        const incoming_symbols = getTickerMarkets(response.active_symbols)
+        const separate_symbols = SeparateSymbols(incoming_symbols)
+        const init_quotes = initiateQuotes(incoming_symbols)
+        setQuotes(init_quotes)
+        setMarkets(separate_symbols)
+
+        if (separate_symbols.close_symbols.length !== 0) {
+            separate_symbols.close_symbols.forEach(symbol => {
+                BinarySocketBase.send(
+                    {
+                        ticks_history: symbol,
+                        end: 'latest',
+                        adjust_start_time: 1,
+                        count: 1,
+                    },
+                    { callback: onCloseSymbolsReceive },
+                )
+            })
+        }
+        BinarySocketBase.send(
+            {
+                ticks: separate_symbols.open_symbols,
+                subscribe: 1,
+            },
+            { callback: onOpenSymbolsReceive },
         )
     }
-    componentWillUnmount() {
-        BinarySocketBase.send({
-            forget_all: 'ticks',
-        })
+    const onOpenSymbolsReceive = response => {
+        console.log(JSON.stringify(quotes))
+        const test = Object.assign({}, quotes)
+        test[response.tick.symbol] = response.tick.quote
+        setQuotes(test)
     }
-    render() {
-        return (
-            <CarouselWapper>
-                {this.state.markets.length === 0 ? null : (
-                    <AutoCarousel
-                        carousel_width="100%"
-                        transition_duration={37000}
-                    >
-                        {this.state.markets.map(symbol => {
-                            return (
-                                <Tick
-                                    key={symbol.symbol}
-                                    display_name={symbol.display_name}
-                                    symbol={symbol.symbol}
-                                    is_exchange_open={!!symbol.exchange_is_open}
-                                    pip={symbol.pip}
-                                ></Tick>
-                            )
-                        })}
-                    </AutoCarousel>
-                )}
-            </CarouselWapper>
-        )
+    const onCloseSymbolsReceive = response => {
+        // setQuotes(
+        //     (quotes[response.echo_req.ticks_history] =
+        //         response.history.prices[0]),
+        // )
     }
+    return (
+        <CarouselWapper>
+            {markets.selected_symbols.length === 0 ? null : (
+                <AutoCarousel carousel_width="100%" transition_duration={37000}>
+                    {markets.selected_symbols.map(symbol => {
+                        return (
+                            <div key={symbol.symbol}>
+                                {symbol.display_name}: {quotes[symbol]}
+                            </div>
+                        )
+                    })}
+                </AutoCarousel>
+            )}
+        </CarouselWapper>
+    )
 }
 
+// export function useQuotes(init = {}) {
+//     const [quotes, setQuotes] = useState(init)
+//     return [quotes, setQuotes]
+// }
+// function useMarkets(
+//     init = {
+//         selected_symbols: [],
+//         close_symbols: [],
+//         open_symbols: [],
+//     },
+// ) {
+//     const [markets, setMarkets] = useState(init)
+//     return [markets, setMarkets]
+// }
 export default Ticker
